relocate/remove unused funcs from Base_GP
