+++ karoo_gp/__init__.py
@@ -1,4 +1,6 @@
 from .pause import pause
-from .base_class import Base_GP, Population, Tree, fx_karoo_pause_refer
+from .tree import Tree
+from .population import Population, fx_fitness_labels_map_maker, fx_fitness_eval
+from .base_class import Base_GP
 
 __version__ = "2.4.0"
